Adjust dock menu settings: ifdef to MAC and remove menu separator
Ellen reports in issue #10 that setAsDockMenu fails to compile on some
other systems, which isn't how this is supposed to work, so we throw it
behind a platform specific ifdef, which is the first in chimery's code.
But now it should compile out of box on Linux.

Additionally, we also remove the separator, which is not super necessary
and also looks bad on the dock icon, where the "About Chimery" item is
relocated automatically by Qt to the App menu and the errant separator
stays in the dock menu, separating nothing.  (In the regular menus the
the separators are automatically removed.)  This UI element is
unneccesary anyway, so let's just remove it.

Fixes #10.